pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepgraph-rag"
version = "2.0.0"
description = "Graph RAG for intersection queries. Find things that HAVE X FOR Y."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "DeepGraph Team"}
]
keywords = [
    "rag",
    "graph-rag",
    "retrieval-augmented-generation",
    "knowledge-graph",
    "neo4j",
    "llm",
    "openai",
    "product-recommendations",
    "research-discovery",
    "multi-hop-reasoning",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "neo4j>=5.16.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
]
ui = [
    "streamlit>=1.31.0",
    "pyvis>=0.3.2",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "streamlit>=1.31.0",
    "pyvis>=0.3.2",
    "pytest>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/deepgraph-rag"
Documentation = "https://github.com/yourusername/deepgraph-rag#readme"
Repository = "https://github.com/yourusername/deepgraph-rag.git"
Issues = "https://github.com/yourusername/deepgraph-rag/issues"

[tool.setuptools.packages.find]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
